requirements.txt
azure-ai-formrecognizer>=3.3.0
azure-core>=1.29.0
python-dotenv>=1.0.0